# api/openapi.yaml
openapi: 3.1.0
info:
  title: Unique Clicks API
  version: 1.0.0
  description: Counts one click per user, author and UTC day.
servers:
  - url: http://localhost:8080
paths:
  /v1/clicks:
    post:
      operationId: trackClick
      summary: Record a unique click
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClickRequest'
      responses:
        '202':
          description: Click accepted and persisted
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          description: Storage operation timed out
...
components:
  schemas:
    ClickRequest:
      type: object
      additionalProperties: false
      required: [user_id, author_id]
      properties:
        user_id:
          type: integer
          format: int64
          minimum: 1
        author_id:
          type: integer
          format: int64
          minimum: 1
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'